+++ comutils/CMakeLists.txt
@@ -3,6 +3,11 @@ file (GLOB devconsole_SRC
 	"devconsole/*.cpp"
 	"devconsole/exceptions/*.hpp"
 	"devconsole/exceptions/*.cpp"
-) 
-add_library (comutils ${devconsole_SRC})
+)
+
+file (GLOB network_SRC
+	"network/*.hpp"
+	"network/*.cpp"
+)
+add_library (comutils ${devconsole_SRC} ${network_SRC})
 target_include_directories (comutils PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
