+++ comutils/network/message.hpp
